requirements.txt
attrs
typing
pytest
sphinx
rinohtype
sphinx_rtd_theme
hydra-core
omegaconf
